refactor(core): several change in components
- modified App.js.
- refactor MyButton.
- refactor Title.
- create StyleMyButton, StyleSigUp.
- add colors to Colors.js
