+++ app-vuelos/src/components/Title.js
@@ -1,5 +1,6 @@
 import { StyleSheet, Text } from "react-native";
 import React from "react";
+import { Colors } from "../theme/Colors";
 
 const Title = ({ titleContent }) => {
   return <Text style={styles.title}>{titleContent}</Text>;
@@ -9,7 +10,7 @@ export default Title;
 
 const styles = StyleSheet.create({
   title: {
-    color: "#5C6EF8",
+    color: Colors.blue,
     fontWeight: "bold",
     fontSize: 25,
     marginBottom: 13
